update 2 year old code :3. moved build to mingw, both code folders cleaned up.
