+++ lsanotif/reg.cpp
